Convert tests in rt_gkgeom into unit tests
